Refactor project structure and Dockerfile for recipes, adding a new REST API service with CRUD operations for recipes. Restructured directories and files, moved to more organized locations, and improved Dockerfile for easier deployment. No new functionalities in the API, just a reorganization for better maintenance and scalability.
