--- post_gin.json
@@ -1,14 +1,0 @@
-{
-  "name": "Ham and cheese toasties",
-  "ingredients": [
-    {
-      "name": "bread"
-    },
-    {
-      "name": "ham"
-    },
-    {
-      "name": "cheese"
-    }
-  ]
-}
