+++ recipes/post_gin.json
@@ -1,0 +1,14 @@
+{
+  "name": "Ham and cheese toasties",
+  "ingredients": [
+    {
+      "name": "bread"
+    },
+    {
+      "name": "ham"
+    },
+    {
+      "name": "cheese"
+    }
+  ]
+}
