+++ recipes/test_data/ham_and_cheese_recipe.json
@@ -1,0 +1,12 @@
+{
+    "name": "Ham and cheese toasties",
+    "ingredients": [
+      {
+        "name": "bread"
+      },{
+        "name": "ham"
+      },{
+        "name": "cheese"
+      }
+    ]
+  }
