+++ restourants/cmd/standart/tmp/build-errors.log
@@ -1,0 +1,1 @@
+exit status 1exit status 1
