--- test_data/ham_and_cheese_with_butter_recipe.json
@@ -1,14 +1,0 @@
-{
-    "name": "Ham and cheese toasties",
-    "ingredients": [
-      {
-        "name": "bread"
-      },{
-        "name": "ham"
-      },{
-        "name": "cheese"
-      },{
-        "name": "butter"
-      }
-    ]
-  }
